單筆: 修正JOIN_STEP，"號之之號"比對邏輯
